Minimum Dart SDK version changed to 2.6.0
flutter_data generated code was causing the analyzer to complain "This
requires the 'extension-methods' experiment to be enabled.
Try enabling this experiment by adding it to the command line when
compiling and running.", as per dart-lang/sdk#37913 . This commit fixes
this by updating the minimum version to one that has said experiment
enabled.